mori-app/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
pydantic>=2.0.0
sse-starlette>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0
structlog>=24.0.0
pyyaml>=6.0.0
